+++ src/PrestControl.Views/PanelView.xaml.cs
@@ -1,0 +1,8 @@
+using System.Windows.Controls;
+
+namespace PrestControl.Views;
+
+public partial class PanelView : UserControl
+{
+    public PanelView() => InitializeComponent();
+}